+++ packages/scenes/src/components/VizPanel/VizPanelRenderer.tsx
@@ -23,10 +23,10 @@ export function VizPanelRenderer({ model }: SceneComponentProps<VizPanel>) {
     placement,
     displayMode,
     hoverHeader,
+    menu,
   } = model.useState();
   const [ref, { width, height }] = useMeasure();
   const plugin = model.getPlugin();
-  const { data } = sceneGraph.getData(model).useState();
   const parentLayout = sceneGraph.getLayout(model);
 
   // If parent has enabled dragging and we have not explicitly disabled it then dragging is enabled
@@ -39,6 +39,9 @@ export function VizPanelRenderer({ model }: SceneComponentProps<VizPanel>) {
 
   // Not sure we need to subscribe to this state
   const timeZone = sceneGraph.getTimeRange(model).state.timeZone;
+
+  // Subscribe to data and apply field overrides
+  const { data } = sceneGraph.getData(model).useState();
   const dataWithOverrides = useFieldOverrides(plugin, fieldConfig, data, timeZone, theme, model.interpolate);
 
   if (pluginLoadError) {
@@ -67,6 +70,11 @@ export function VizPanelRenderer({ model }: SceneComponentProps<VizPanel>) {
     titleItems.push(<model.state.$timeRange.Component model={model.state.$timeRange} />);
   }
 
+  let panelMenu;
+  if (menu) {
+    panelMenu = <menu.Component model={menu} />;
+  }
+
   return (
     <div ref={ref as RefCallback<HTMLDivElement>} style={{ position: 'absolute', width: '100%', height: '100%' }}>
       <PanelChrome
@@ -81,6 +89,7 @@ export function VizPanelRenderer({ model }: SceneComponentProps<VizPanel>) {
         titleItems={titleItems}
         dragClass={dragClass}
         dragClassCancel={dragClassCancel}
+        menu={panelMenu}
       >
         {(innerWidth, innerHeight) => (
           <>