+++ packages/scenes/src/core/types.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import { MonoTypeOperatorFunction, Observer, SubscriptionLike, Unsubscribable } from 'rxjs';
+import { MonoTypeOperatorFunction, Unsubscribable } from 'rxjs';
 
 import {
   BusEvent,
@@ -69,7 +69,7 @@ export interface SceneObject<TState extends SceneObjectState = SceneObjectState>
   readonly urlSync?: SceneObjectUrlSyncHandler<TState>;
 
   /** Subscribe to state changes */
-  subscribeToState(observer?: Partial<Observer<TState>>): SubscriptionLike;
+  subscribeToState(handler: SceneStateChangedHandler<TState>): Unsubscribable;
 
   /** Subscribe to a scene event */
   subscribeToEvent<T extends BusEvent>(typeFilter: BusEventType<T>, handler: BusEventHandler<T>): Unsubscribable;
@@ -83,10 +83,15 @@ export interface SceneObject<TState extends SceneObjectState = SceneObjectState>
   /** How to modify state */
   setState(state: Partial<TState>): void;
 
-  /** Called when the Component is mounted. A place to register event listeners add subscribe to state changes */
+  /**
+   * Called when the Component is mounted. This will also activate any $data, $variables or $timeRange scene object on this level.
+   * Don't override this in your custom SceneObjects, instead use addActivationHandler from the constructor. The activation handler can return a deactivation handler.
+   **/
   activate(): void;
 
-  /** Called when component unmounts. Unsubscribe and closes all subscriptions  */
+  /** Called when component unmounts. This will also deactivate any $data, $variables or $timeRange scene object on this level.
+   * Don't override this in your custom SceneObjects, instead use addActivationHandler from the constructor. The activation handler can return a deactivation handler.
+   */
   deactivate(): void;
 
   /** Get the scene root */
@@ -103,7 +108,16 @@ export interface SceneObject<TState extends SceneObjectState = SceneObjectState>
 
   /** Force a re-render, should only be needed when variable values change */
   forceRender(): void;
+
+  /**
+   * Allows external code to register code that is executed on activate and deactivate. This allow you
+   * to wire up scene objects that need to respond to state changes in other objects from the outside.
+   **/
+  addActivationHandler(handler: SceneActivationHandler): void;
 }
+
+export type SceneActivationHandler = () => SceneDeactivationHandler | void;
+export type SceneDeactivationHandler = () => void;
 
 export type SceneLayoutChild = SceneObject<SceneLayoutChildState | SceneLayoutState>;
 
@@ -171,6 +185,7 @@ export type SceneObjectUrlValue = string | string[] | undefined | null;
 export type SceneObjectUrlValues = Record<string, SceneObjectUrlValue>;
 
 export type CustomTransformOperator = (context: DataTransformContext) => MonoTypeOperatorFunction<DataFrame[]>;
+export type SceneStateChangedHandler<TState> = (newState: TState, prevState: TState) => void;
 
 export type DeepPartial<T> = {
   [K in keyof T]?: T[K] extends object ? DeepPartial<T[K]> : T[K];