+++ packages/scenes/src/querying/SceneQueryRunner.ts
@@ -1,11 +1,11 @@
 import { cloneDeep } from 'lodash';
 import { mergeMap, MonoTypeOperatorFunction, Unsubscribable, map, of } from 'rxjs';
 
+import { DataQuery, DataSourceRef, LoadingState } from '@grafana/schema';
+
 import {
   CoreApp,
-  DataQuery,
   DataQueryRequest,
-  DataSourceRef,
   DataTransformerConfig,
   PanelData,
   preProcessPanelData,
@@ -28,7 +28,7 @@ import { VariableValueRecorder } from '../variables/VariableValueRecorder';
 let counter = 100;
 
 export function getNextRequestId() {
-  return 'QS' + counter++;
+  return 'SQR' + counter++;
 }
 
 export interface QueryRunnerState extends SceneObjectStatePlain {
@@ -58,23 +58,25 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
       this.onVariableUpdatesCompleted(variables, dependencyChanged),
   });
 
-  public activate() {
-    super.activate();
+  public constructor(initialState: QueryRunnerState) {
+    super(initialState);
 
+    this.addActivationHandler(this._onActivate);
+  }
+
+  private _onActivate = () => {
     const timeRange = sceneGraph.getTimeRange(this);
 
     this._subs.add(
-      timeRange.subscribeToState({
-        next: (timeRange) => {
-          this.runWithTimeRange(timeRange.value);
-        },
+      timeRange.subscribeToState((timeRange) => {
+        this.runWithTimeRange(timeRange.value);
       })
     );
 
     if (this.shouldRunQueriesOnActivate()) {
       this.runQueries();
     }
-  }
+  };
 
   /**
    * Handles some tricky cases where we need to run queries even when they have not changed in case
@@ -169,10 +171,19 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
       this.setState({ isWaitingForVariables: false });
     }
 
-    const { datasource, minInterval, queries } = this.state;
+    const { minInterval, queries } = this.state;
     const sceneObjectScopedVar: Record<string, ScopedVar<SceneQueryRunner>> = {
       __sceneObject: { text: '__sceneObject', value: this },
     };
+
+    // Simple path when no queries exist
+    if (!queries?.length) {
+      this.onDataReceived({
+        state: LoadingState.Done,
+        series: [],
+        timeRange,
+      });
+    }
 
     const request: DataQueryRequest = {
       app: CoreApp.Dashboard,
@@ -190,6 +201,7 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
     };
 
     try {
+      const datasource = this.state.datasource ?? findFirstDatasource(request.targets);
       const ds = await getDataSource(datasource, request.scopedVars);
 
       // Attach the data source name to each query
@@ -228,6 +240,15 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
     const preProcessedData = preProcessPanelData(data, this.state.data);
     this.setState({ data: preProcessedData });
   };
+}
+
+function findFirstDatasource(targets: DataQuery[]): DataSourceRef | undefined {
+  for (const t of targets) {
+    if (t.datasource != null) {
+      return t.datasource;
+    }
+  }
+  return undefined;
 }
 
 export function getTransformationsStream(